refactor(sheaves): drop 'stalk' terminology in favor of 'candidate'
Renames stalk.ts → match.ts and getStalk → getMatchingProviders. The
function returns providers (pre-evaluation), not candidates, so the new
name agrees with the type. Drops the Stalk glossary entry from the
README — every other doc and code comment already used candidate(s) for
both the element and the set.
