+++ packages/sheaves/README.md
@@ -32,25 +32,20 @@ This is an element of the presheaf F = F_sem x F_op.
 > narrower open set.
 
 **Candidate** — An equivalence class of providers at an invocation point,
-identified by metadata. At dispatch time, providers in the stalk with identical
+identified by metadata. At dispatch time, matching providers with identical
 metadata are collapsed into a single candidate; the system picks an arbitrary
 representative for dispatch. If two capabilities are indistinguishable by
 metadata, the sheaf has no data to prefer one over the other.
 
-> Two `getBalance(string)` providers both with `{ cost: 1 }` collapse into
-> one candidate. The policy never sees both — it receives one representative.
+> At `("getBalance", "alice")` the candidate set might contain two entries
+> (cost 1 vs 100); at `("transfer", ...)` it might contain one. Two
+> `getBalance(string)` providers both with `{ cost: 1 }` collapse into one
+> candidate — the policy never sees both, it receives one representative.
 
-**Stalk** — The set of candidates matching a specific `(method, args)` invocation,
-computed at dispatch time by guard filtering and then collapsing equivalent
-entries.
-
-> Stalk at `("getBalance", "alice")` might contain two candidates (cost 1 vs 100);
-> stalk at `("transfer", ...)` might contain one.
-
-**Policy** — An `async function*` coroutine that yields candidates from a
-multi-candidate stalk in preference order. See [POLICY.md](./docs/POLICY.md)
-for the coroutine protocol, `PolicyContext`, and the semantic equivalence
-assumption required of all policies.
+**Policy** — An `async function*` coroutine that yields candidates in
+preference order when more than one matches an invocation. See
+[POLICY.md](./docs/POLICY.md) for the coroutine protocol, `PolicyContext`, and
+the semantic equivalence assumption required of all policies.
 
 At dispatch time, metadata is decomposed into **constraints** (keys with the
 same value across every candidate — topologically determined, not a choice) and
@@ -59,9 +54,9 @@ receives only options on each candidate; constraints arrive separately in the
 context.
 
 > `argmin` by cost, `argmin` by latency, or any custom selection logic. The
-> policy is never invoked when the stalk resolves to a single candidate — either
-> because only one provider matched, or because all matching providers had
-> identical metadata and collapsed to one representative.
+> policy is never invoked when only one candidate remains — either because
+> only one provider matched, or because all matching providers had identical
+> metadata and collapsed to one representative.
 
 **Sheaf** — The authority manager returned by `sheafify`. Holds the provider
 data (frozen at construction time) and exposes factory methods that
@@ -79,13 +74,13 @@ const sheaf = sheafify({ name: 'Wallet', providers });
 At each invocation point `(method, args)` within a granted section:
 
 ```
-getStalk(providers, method, args)    presheaf → stalk (filter by guard)
-evaluateMetadata(stalk, args)        metadata specs → concrete values
-collapseEquivalent(stalk)            locality condition (quotient by metadata)
-decomposeMetadata(collapsed)         restriction map (constraints / options)
-policy(candidates, { method, args,   operational selection (extra-theoretic)
+getMatchingProviders(providers, method, args)  presheaf → matches (filter by guard)
+evaluateMetadata(matches, args)                metadata specs → concrete values
+collapseEquivalent(candidates)                 locality condition (quotient by metadata)
+decomposeMetadata(collapsed)                   restriction map (constraints / options)
+policy(candidates, { method, args,             operational selection (extra-theoretic)
                   constraints })
-dispatch to chosen.exo           evaluation
+dispatch to chosen.exo                         evaluation
 ```
 
 The pipeline short-circuits at two points: if only one provider matches the
@@ -93,7 +88,7 @@ guard, it is invoked directly without evaluate/collapse/policy; if all matching
 providers collapse to an identical candidate, the single representative is invoked
 without calling the policy.
 
-`callable` metadata specs make the stalk shape depend on the invocation
+`callable` metadata specs make the candidate set depend on the invocation
 arguments. A `swap(amount)` provider can produce `{ cost: 'low' }` for small
 amounts and `{ cost: 'high' }` for large ones, yielding a different set of
 candidates — and potentially a different policy outcome — for the same method
@@ -108,14 +103,14 @@ interchangeable. Under the sheaf condition (effect-equivalence), this recovers
 the classical equivalence relation on germs.
 
 **Pseudosheafification.** The sheafification functor would precompute the full
-etale space. This system defers to invocation time: compute the stalk,
-collapse, decompose, select via policy. The trade-off is that global coherence
-(a policy choosing consistently across points) is not guaranteed.
+etale space. This system defers to invocation time: match by guard, evaluate
+metadata, collapse, decompose, select via policy. The trade-off is that global
+coherence (a policy choosing consistently across points) is not guaranteed.
 
 **Restriction and gluing are implicit.** Guard restriction induces a
 restriction map on metadata: restricting to a point filters the presheaf to
-covering providers (`getStalk`), then `decomposeMetadata` strips the metadata
-to distinguishing keys — the restricted metadata over that point. The join
+covering providers (`getMatchingProviders`), then `decomposeMetadata` strips
+the metadata to distinguishing keys — the restricted metadata over that point. The join
 works dually: the union of two providers has the join of their metadata, and
 restriction at any point recovers the local distinguishing keys in O(n).
 Gluing follows: compatible providers (equal metadata on their overlap) produce a
